requirements.txt
asgiref==3.5.2
celery==5.3.4
django==4.1
requests==2.28.1
django-redis==5.2.0.1
djangorestframework==3.13
gunicorn==20.1.0